Fix the type vars fixup phase for unirecords
