+++ src/parser/uniterm.rs
@@ -5,11 +5,11 @@ use utils::{build_record, FieldPathElem};
 
 use crate::{
     position::{RawSpan, TermPos},
-    term::{MergePriority, MetaValue, RecordAttrs, RichTerm, Term},
+    term::{Contract, MergePriority, MetaValue, RecordAttrs, RichTerm, SharedTerm, Term},
     types::{AbsType, Types, UnboundTypeVariableError},
 };
 
-use std::convert::TryFrom;
+use std::{collections::HashSet, convert::TryFrom};
 
 /// A node of the uniterm AST. We only define new variants for constructs common to types and
 /// terms. Otherwise, we piggyback on the existing ASTs to avoid duplicating code and definitions.
@@ -215,7 +215,10 @@ impl TryFrom<UniRecord> for RichTerm {
     /// Convert a `UniRecord` to a term. If the `UniRecord` has a tail, it is first interpreted as
     /// a type and then converted to a contract. Otherwise it is interpreted as a record directly.
     /// Fail if the `UniRecord` has a tail but isn't syntactically a record type either.
-    fn try_from(ur: UniRecord) -> Result<Self, ParseError> {
+    ///
+    /// We also fix the type variables of the type appearing inside annotations (see
+    /// [`fix_type_vars`]).
+    fn try_from(mut ur: UniRecord) -> Result<Self, ParseError> {
         let pos = ur.pos;
 
         let result = if let Some((_, tail_pos)) = ur.tail {
@@ -224,12 +227,14 @@ impl TryFrom<UniRecord> for RichTerm {
                 .map_err(|InvalidRecordTypeError(pos)| {
                     ParseError::InvalidUniRecord(pos.unwrap(), tail_pos.unwrap(), pos.unwrap())
                 })
-                .and_then(|ty| {
+                .and_then(|mut ty| {
+                    fix_type_vars(&mut ty);
                     ty.contract().map_err(|UnboundTypeVariableError(id)| {
                         ParseError::UnboundTypeVariables(vec![id], pos.unwrap())
                     })
                 })
         } else {
+            fix_fields_types(ur.fields.iter_mut().map(|(_, rt)| rt));
             Ok(RichTerm::from(build_record(
                 ur.fields.into_iter(),
                 ur.attrs,
@@ -263,6 +268,101 @@ impl TryFrom<UniRecord> for Types {
             ur.clone()
                 .into_type_strict()
                 .or_else(|_| RichTerm::try_from(ur).map(|rt| Types(AbsType::Flat(rt))))
+        }
+    }
+}
+
+/// Post-process a type at the right hand side of an annotation by replacing unbound type variables
+/// by the same variables but seen as custom contracts (`AbsType::Var(id)` to
+/// `AbsType::Flat(Term::Var(id))`).
+///
+/// Since parsing is done bottom-up, and given the specification of the uniterm syntax for
+/// variables occurring in types, we can't always know right away if a such a variable occurrence
+/// is actually a type variable or a term variable seen as a custom contract.
+///
+/// Take for example `a -> b`. At this stage, `a` and `b` could be both variables referring to a
+/// contract (e.g. in `x | a -> b`) or a type variable (e.g. in `x | forall a b. a -> b`),
+/// depending on enclosing `forall`s. To handle both cases, we initially parse all variables inside
+/// types as type variables. When reaching the right-hand side of an annotation, because `forall`s
+/// can only bind locally in a type, we can then decide the actual nature of each occurrence. We
+/// thus recurse into the newly constructed type to change those type variables that are not
+/// actually bound by a `forall` to be term variables. This is the role of `fix_type_vars()`.
+///
+/// Once again because `forall`s only bind variables locally, and don't bind inside contracts, we
+/// don't have to recurse into contracts and this pass will only visit each node of the AST at most
+/// once in total.
+///
+/// There is one subtlety with unirecords, though. A unirecord can still be in interpreted as a
+/// record type later. Take the following example:
+///
+/// ```
+/// let mk_pair : forall a b. a -> b -> {fst: a, snd: b} = <exp>
+/// ```
+///
+/// Since this unirecord will eventually be interpreted as a record type, we can't know yet when
+/// parsing `fst: a` if `a` will be a type variable or a term variable (while, for all other
+/// constructs, an annotation is a boundary that `forall` binders can't cross). In this example,
+/// there is indeed an enclosing forall binding `a`. With unirecords, before fixing type variables,
+/// we have wait until we eventually convert the unirecord to a term (in which case we fix all the
+/// top-level annotations) or a type (in which case we do nothing: the enclosing type will trigger
+/// the fix once it's fully constructed). Fixing a unirecord prior to a conversion to a term is
+/// done by [`fix_fields_types`].
+pub fn fix_type_vars(ty: &mut Types) {
+    fn fix_type_vars_aux(ty: &mut Types, mut bound_vars: HashSet<Ident>) {
+        match ty.0 {
+            AbsType::Dyn()
+            | AbsType::Num()
+            | AbsType::Bool()
+            | AbsType::Str()
+            | AbsType::Sym()
+            | AbsType::Flat(_)
+            | AbsType::RowEmpty() => (),
+            AbsType::Arrow(ref mut s, ref mut t) => {
+                fix_type_vars_aux(s.as_mut(), bound_vars.clone());
+                fix_type_vars_aux(t.as_mut(), bound_vars.clone());
+            }
+            AbsType::Var(ref mut id) => {
+                if !bound_vars.contains(id) {
+                    let id = std::mem::take(id);
+                    let pos = id.pos;
+                    ty.0 = AbsType::Flat(RichTerm::new(Term::Var(id), pos));
+                }
+            }
+            AbsType::Forall(ref id, ref mut ty) => {
+                bound_vars.insert(id.clone());
+                fix_type_vars_aux(&mut *ty, bound_vars);
+            }
+            AbsType::RowExtend(_, ref mut ty_opt, ref mut tail) => {
+                (*ty_opt)
+                    .iter_mut()
+                    .for_each(|ty| fix_type_vars_aux(ty.as_mut(), bound_vars.clone()));
+                fix_type_vars_aux(tail.as_mut(), bound_vars);
+            }
+            AbsType::DynRecord(ref mut ty)
+            | AbsType::List(ref mut ty)
+            | AbsType::Enum(ref mut ty)
+            | AbsType::StaticRecord(ref mut ty) => fix_type_vars_aux(ty.as_mut(), bound_vars),
+        }
+    }
+
+    fix_type_vars_aux(ty, HashSet::new())
+}
+
+/// Fix the type variables of the types appearing as annotations of the fields of a record. See
+/// [`fix_type_vars`].
+pub fn fix_fields_types<'a, I: Iterator<Item = &'a mut RichTerm>>(fields: I) {
+    for rt in fields {
+        match SharedTerm::make_mut(&mut rt.term) {
+            Term::MetaValue(ref mut m) => {
+                if let Some(Contract { ref mut types, .. }) = m.types {
+                    fix_type_vars(types);
+                }
+
+                for ctr in m.contracts.iter_mut() {
+                    fix_type_vars(&mut ctr.types);
+                }
+            }
+            _ => (),
         }
     }
 }
